user login and session

+++ controllers/cats_controller.rb
@@ -7,14 +7,17 @@ class CatsController < ControllerBase
   end
 
   def new
+    return redirect_to "/cats" unless current_user
     render "new"
   end
 
   def create
+    return redirect_to "/cats" unless current_user
     @cat = Cat.new(params["cat"])
+    @cat.set(human_id: current_user.id)
     if @cat.save
       flash["flash now"] = "#{@cat.name} created!"
-      redirect_to("/cats")
+      redirect_to "/cats"
     else
       flash.now["flash now"] = "Error!"
       render "new"
@@ -28,12 +31,14 @@ class CatsController < ControllerBase
 
   def edit
     @cat = Cat.find(params[:id])
+    return redirect_to "/cats" unless current_user && current_user.id == @cat.human_id
     render "edit"
   end
 
   def update
     @cat = Cat.find(params[:id])
-    @cat.set(params["cat"])
+    return redirect_to "/cats" unless current_user && current_user.id == @cat.human_id
+    @cat.set(name: params["cat"]["name"])
     if @cat.save
       redirect_to("/cats/#{@cat.id}")
     else
@@ -45,6 +50,7 @@ class CatsController < ControllerBase
 
   def destroy
     cat = Cat.find(params[:id])
+    return redirect_to "/cats" unless current_user && current_user.id == cat.human_id
     cat.destroy
     flash["flash now"] = "#{cat.name} woke up!"
     redirect_to("/cats")
